Move position modified to separate file
